requirements.txt
spacy
bs4
urllib3
nltk
spacy-transformers